baseline/requirements.txt
openai
python-dotenv
chromadb
langchain-chroma
langchain-community
langchain